Add content to pesto_pasta.html
I created a new directory in the odin-recipes directory called "images", and then moved the image Pesto_Pasta_image.webp into it. In pesto-pasta.html I added the image of the pasta using the <image> tag. Then, I added an unordered list of ingredients using the <ul> and <li> tages and a ordered list with steps for cooking using the <ol> and <li> tags.